Delete button no longer crashes.

Related work items: #131

+++ 8-Bit-Twist/8-Bit-Twist/Pages/Admin/Products/Manage.cshtml.cs
@@ -79,6 +79,10 @@ namespace _8_Bit_Twist.Pages.Admin
 
         public async Task<IActionResult> OnPostDelete()
         {
+            if (!await _invManager.ProductExists(ID.GetValueOrDefault()))
+            {
+                return RedirectToPage("/Admin/Products/Index");
+            }
             await _invManager.DeleteProduct(ID.GetValueOrDefault());
             return RedirectToPage("/Admin/Products/Index");
         }
